requirements.txt
Django==5.0.2
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
mysqlclient==2.2.4
python-dotenv==1.0.1
django-taggit==5.0.1 